requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pycryptodome
httpx